backend/ml_pipeline/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
notebook
ipykernel
pyyaml
